Fiddling about
Trying to get a couple of tests to work (failing because of some weird error where the `vtimezone` component of the calendar is losing its `$objName` field value for unknown reasons): `RetainCalendarHistoryTest.testInstantiationFromCalendar` and `RetainCalendarHistoryTest.testInstantiationFromRetainCalendarHistory`. The corresponding test for a `Calendar` in `FilteredCalendarTest` seems fine, but I'm wondering if there's something about multiple passes through a calendar that's messing things up.

Meanmwhile, also trying to clean up a few errors Scrutinizer identified.